Consider move to higher level for value object and interface.
