+++ docs/HtmlStringable.md
@@ -18,8 +18,9 @@ $this->Form->postLink($icon, ['action' => 'delete', $id]);
 ```
 
 Use can also use the `Html` value object directly:
+
 ```php
-use Templating\View\Html\Html;
+use Templating\View\Html;
 
 $html = Html::create('<i>text</i>');
 $this->Html->link($html, '/my/url');
@@ -28,7 +29,7 @@ $this->Html->link($html, '/my/url');
 The same goes for any custom snippet of yours, e.g.
 
 ```php
-use Templating\View\Html\HtmlStringable;
+use Templating\View\HtmlStringable;
 
 class SvgGraph implements HtmlStringable { ... }
