+++ src/View/HtmlStringable.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Templating\View\Html;
+namespace Templating\View;
 
 use Stringable;
